// src/enums/badge/badge-type.enum.ts
export enum BadgeType {
  FIRST_AUTHOR = 'first_author',
  COMMENT_GURU = 'comment_guru',
  POPULAR_AUTHOR = 'popular_author',
  TOTAL_LIKES = 'total_likes',
  TOPIC_MASTER = 'topic_master',
  EDITORS_CHOICE = 'editors_choice',
  IDEA_GENERATOR = 'idea_generator',
  COMMUNITY_HEART = 'community_heart',
  KNOW_IT_ALL = 'know_it_all',
}